Fix .NET Framework 4.8 compatibility issues in Library and Tests

+++ Library/System/Console.Unix.cs
@@ -7,9 +7,9 @@ public static partial class ConsoleExtensions
     private static SafeFileHandle GetStdHandle(int handleType)
         => handleType switch
         {
-            Interop.Kernel32.HandleTypes.STD_INPUT_HANDLE => new(0, false),
-            Interop.Kernel32.HandleTypes.STD_OUTPUT_HANDLE => new(1, false),
-            Interop.Kernel32.HandleTypes.STD_ERROR_HANDLE => new(2, false),
+            Interop.Kernel32.HandleTypes.STD_INPUT_HANDLE => new((IntPtr)0, false),
+            Interop.Kernel32.HandleTypes.STD_OUTPUT_HANDLE => new((IntPtr)1, false),
+            Interop.Kernel32.HandleTypes.STD_ERROR_HANDLE => new((IntPtr)2, false),
             _ => throw new ArgumentOutOfRangeException(nameof(handleType), "Invalid standard handle type.")
         };
 }
